Added test for [#4293] define something and use it
This test appears to only come up in agent/server
mode.  I could not trigger the bug in 2.6.0rc4
using stand alone apply mode.

This test doesn't fail on any of the 2.6.0
release candidates, as far as I can tell.  I'm
not sure if the bug and fix were transitional
between release candidates or not...

Updated list of issues needing tests.

Signed-off-by: Jeff McCune <jeff@puppetlabs.com>