Fix the cron removal test to be compatible with Ubuntu
Ubuntu stores it's cron jobs in a different place than Redhat, so the
test written to pass on Redhat failed on Ubuntu.  I used facter's
operatingsystme fact to branch in the test, which I see has been done in
one other test so far.  More of this will certainly be necessary in the
future as we move to more cross platform testing, so we'll have to
develop best practices as we move into that area.

Also cleaned up the test to take advantage of Jesse's new cleanup
method.

Reviewed by: Nick Lewis