Added spec test for [#4404] Stage[main] on left
This spec test fails with puppet 2.6.0 and 2.6.1rc1.

It should exit 0 when the issue is fixed.

The script runs with set -e and will properly exit
with status code 10 ($EXIT_FAILURE) if puppet
does not exit successfully.  All other exit codes
other than 0 and 10 should be considered an error
and not a failure of the specification.

Signed-off-by: Jeff McCune <jeff@puppetlabs.com>