Fix the cron tests to work with Ubuntu
I fixed one of these cron tests in an earlier commit, so this is making
the same change of taking into account that Ubuntu stores it's cron
entries in a different directory.

The should_update_existing_spec still fails on Ubuntu, but now not
because it's looking in the wrong directory.  More research to determine
why it's still failing.