Make the test failure output more readable
When the test failure output only had the failing test name it made
sense to have the lines numbered.  Now that the test error messages are
also in the failure list, the line numbering is confusing and hard to
read.  This gets rid of line numbering and replaces it with test
numbering, spaces out the failure results and moves the summary to the
end.