Refactor en route to #5505 -- lazy require of tests
Moves the "require" of test files from the directory search to right before the
tests are run.  This commit changes the behavior in a way that should have no
user-visible consequence (if it does, some test(s) are so pathalogical/sneaky
that we have worse things to worry about, such as how those tests got in).