Fix for #5546 -- Pull common test idioms into a DSL
This is a large commit in terms of the number of lines touched, but isn't
conceptually complicated.  A half dozen patterns were identifed in the
existing tests, these were added as methods to the test wrapper, and then
all occurences of them were replaced with calls to the methods.  The main
reason for doing it all as one big commit is that several of the patterns
"fit together" in a certain way and it was easier to change them at the same
time then to erect scafolding.