Fix for #5562 -- give hosts properties with defaults
This changes hosts from strings (their names) to objects with properties and
uses the config['CONFIG'] values as the defaults.  It not only simplifies some
of the code (e.g. determing the roles of a host, or the number of files to
serve) but also enables many things, such as having different versions of
puppet/ruby/whatever on each host.