Improve the DSL for remote puppet and facter invocation
Invoking these tools means, at least in my test environment, configuring the
RUBYLIB and PATH environment variables before invoking the executable.

This change implements that, and centralizes the logic for building those in
the 'puppet' and 'facter' DSL methods, which also makes them nicer to invoke...

For now the environment for doing that comes from the configuration file, statically, but in future it should be configured on a per-host basis.