Cache the SSH connection for the file of a host object.
This allows us to reuse the authenticated SSH connection to the target for
each remote command execution.  This saves me between 10 and 15 percent of
overall test execution time on my test systems - which are very fast and have
no other significant load, so real-world results should only be better than
that.