Unify SSH options and access through the cache.
This brings the SSH options to the top level, and allows the user to override
them with the configuration file - sensible defaults are used when there is
nothing given, and I expect most all our users to use them.

It also unifies the scp operations through the same SSH connection, invoking
the subsystem on a new channel dynamically.  In the process the action class
is simplified and made easier to deal with.