Eliminate the "action" class entirely.
Now we have the DSL to encapsulate the actions, and SSH management centralized
in the host class, the action infrastructure no longer serves a significant
purpose.

By eliminating it we can simplify the overall structure of the tests and make
it easier to see where the DSL and framework hang together.