Simple but unsatisfying fix for #5671 -- specifiy non-root users
The options hook that I was going to use to implement a by-call-with-defaults
specification of the user went away in 68835137a7bf52ae879b914a1251413c26745769
so this just makes it global to the host.  Ultimately, the global to the host
version should be a default which could be overridden on any given call.