Feature #5519 -- support installing from both git and PE
The test harness now supports arbitrary (but verified) modes of installation,
to support PE and git installation.  This change implements the infrastructure
for that, and migrates the existing code to perform the PE installation.

We now support a suite of 'setup' changes - that use the same infrastructure
as running tests - that is used to implement the process.  Installation modes
are created under 'setup/${type}', and stuff from 'setup/common' is always run
after the base installation.

Finally, we also expose the 'options' parsed from the command line to the
tests, because we need to interact with those when doing the git install;
tests can't add command line options, but they can inspect them.