(#6163) Make a log dir for the test output
It will be easier to grep and tail the output in the logs if they're in
their own directory.  Long term we should also a way for the test output
to go to the screen as well, but for now a simple log dir is a big
improvement.

Git won't let me add an empty directory, so I followed the trickery
mentioned here:

http://help.github.com/git-cheat-sheets/

to allow me to add a log dir where the contents will be ignored by git.