maint: Put option parsing in a reusable class
Currently the only way to access the options was to reparse them, and if
you did this from certain points in the harness that didn't have access
to the command line args, you just got the defaults.

This change puts the options parsing in a class that stores the options
once parsed so that you don't have to reparse them at later points in
the test harness.

Paired-with: Dominic Maraglia