Fixed logic error in gather git sha1
Harness clones puppet and facter on the first host in the config file
that may or may not be the Master.  Next step, the harness queries
the Master node for puppet/facter sha1; however, if Master node was
not the first definded node in the condfig file, the Master will
not have had the puppet/facter code checkout, hence triggering a
failure.