(#6688) Eliminate global $workdir from the suite.
We used to play around with our working directory during the suite run, but
that turned out to be troublesome.  This commit just eliminates the last
remnants of that activity.

We also eliminate the chdir back at the end of the suite, which would have had
no practical process effect – but was kind of confusing.