(#6688) Push the log instance into the test suite...
The log interactions were only ever part of the test suite proper; push the
entire instance down, including summary dumping and other finalization.

This makes it easier to hook in a standard finalizer for logs, which we are
going to need to write out JUnit XML   which foolishly puts summary data into
the very top of the test data.