(#6688) Merge the "instance" parts of log into the test suite.
The log was used for two purposes: as a class, to contain our logging helpers,
and as an instance to allow the test suite to accumulate state during the run.

This merges all the instance stuff out into the suite, and just uses it there
quietly.  This is about 99 percent clean, but required a bit of hacking
because we have more instances of TestSuite than we did of log.