(#6688) More setup, less hard-coded logic, for VM management.
The mrproper and vmrun features were hard-coded into the systest.rb file, when
we already had suitable infrastructure for running them as a regular part of
the test process.

They are now merged down into setup stages that, if the options are set, act
just like any other part of the setup infrastructure.  This brings
consistency, even if there is now more ugly between options and the setup
"test" cases.