(#6688) Factor common test suite behaviour to the suite class.
This moves the behaviour of "run, and succeed or abort the harness" out to a
central location in the TestSuite class.  This moves the common behaviour to
the common location.

I am unconvinced by the use of 'exit' here as our abort, but it mirrors the
existing behaviour of the code, so I preserve it.  At least we only have one
place to visit to change it now...