(#6688) Remove std{out,err} manipulation from the entire codebase.
We now arrange for our logging tools to write to their configured destinations
appropriately.  This avoids the need to futz around with normal print routines
during running of the test code.

This does introduce, or rather reveal, some strange global test state during
the routine operation of the test tools, since printing a summary leaks the
summary log as the target.

This is actually the way things were before, but is now a bit more clear.  We
actually work around that to make it less shocking in future...