(#6688) Option to disable ANSI colour in logs.
This is a new CLI option, which has the added benefit that we don't have to
handle injection of ANSI escape sequences into the CI server logging, or more
importantly, into the JUnit XML output.

Currently this is done during message output, but this also opens the door to
pushing the output down into output drivers at a later date.