(#6688) Implement JUnit summary output from the test suite.
This uses REXML to generate the output in the appropriate format, along with
some manual escaping to work around embedded ANSI colour codes from the Puppet
tools run in the suite.

This targets, specifically, Hudson-compatible output, so focuses on the data
it will correctly present rather than aiming for a more general purpose
consumer.