maint: Add a wrapper method for running a master
with_master_running_on(host) can be used to run a master for the duration of
the passed block. This method will handle starting a master, waiting for it to
be ready for connections, and then later taking it down and waiting for it to
be completely shut down.

with_master_running_on(master) do
  run_agent_on(agents)
end

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>
Reviewed-By: Dominic Maraglia <dominic@puppetlabs.com>