(#11618) Add root SSH key synchronization setup step
Without this patch the target systems under test may have out of date
authorized key files since they're usually VM snapshots.  This is
problem if there is a test problem requiring investigating by technical
staff.

This patch fixes the problem by synchronizing the
~root/.ssh/authorized_keys file in the setup early phase.  The behavior
happens by default on all systems.  We assume it will work and intended
to come back and filter out target nodes that are known not to work.

A command line option, --no-root-keys is provided to disable this change
in behavior.

Reviewed-by: Dominic Maraglia