(#11621) Add --dhcp-renew option to enable lease renewal
Without this patch the DHCP lease renewal step of the early setup phase
is always executed.

This is a problem on my CentOS 6 VMware Fusion VM because the interface
is brought down, disconnecting the SUT from the network before the ifup
command can be issued.  This leaves the target system offline and
subsequent steps timeout.

This patch works around the problem by providing and documenting the
--dhcp-renew option to disable renewals.

Reviewed-by: Dominic Maraglia