(#11767) Sync ntp on Windows
By default, the acceptance tests perform an ntp sync. On Windows 2003
and up, this can be done using the time service `w32time`, not to be
confused with the command line utility `w32tm.exe` that interacts with
the service.

If the w32time service is already running, then `net start` will
return exit code 2.