(#11767) Skip creating puppet user and group on Windows
The `getent`, `useradd`, and `groupadd` commands do not exist on
Windows (nor Cygwin), so we skip creating the puppet user and group on
Windows.

Ideally, we should only create the puppet user and group in a `master`
role (when we need to switch uid). However, due to bug #9862 the puppet
group is required on Unix, even for `puppet agent` and `puppet apply`.

Since #9862 doesn't apply on Windows, we can safely skip creating the
puppet user and group on Windows.