(#11767) Move defaults into type-specific hosts
Previously, the acceptance tests were adding POSIX default paths to
the PATH environment variable passed to Windows hosts, e.g. /usr/bin,
which happens to correspond to the location of the Cygwin utilities
that we don't want the tests to use on Windows.

This commit creates host subclasses that contain the default paths that
are applicable for that platform. It also adds a `Host.create` factory
method that knows how to create the appropriate type of `Host` instance
based on the platform specified in the configuration file.

If we need to support PE-specific install locations on Windows, then it
will be easy to change (in one place).