(#11767) Don't create a new SSH connection for each TestCase
Previously, a new Host instance was being created for each combination
of TestSuite, TestCase, and Host. As a result, multiple SSH
connections, including SSH authentication, were being performed
unnecessarily. This was actually a regression introduced in
b62f8056f0b5499ad9341487c1fca6f328c0830b.

Given a sufficiently large number of TestCases, the test driver and/or
system under test could run out of file descriptors.

This commit moves the creation of hosts to the top-level, so that the
same SSH connection can be reused across multiple TestSuites and
TestCases. In other words, the acceptance tests will make at most one
SSH connection per host.

This commit also adds a method for closing the SSH connection (if one
was created) and explicitly closes the connection when the tests are
done.