(#11767) Make RemoteExec commands work on Windows
Previously, the acceptance tests could not launch puppet and facter via
SSH and Cygwin on the Windows side, due to Cygwin's bash shell
attempting to invoke ruby.exe on `C:\Ruby187\bin\puppet` and failing,
because that file contains Windows-style paths that confuse bash.

This commit simply wraps the puppet and facter commands within a
`cmd.exe` shell to escape from the bash shell (though the environment
is inherited).