(#11767) Don't assume host['puppetpath'] is /etc/puppet
Previously, the acceptance tests failed on Windows because puppet.conf
does not live in /etc/puppet on Windows.

This commit changes the default value of `puppetpath` for Windows
hosts relative to the CommonAppData folder. Sadly, the location of
this folder is different between 2003 and 2008, and the folder usually
contains spaces, e.g. C:\Documents and Settings\All Users\Application
Data.

This commit uses the Cygwin utility `cygpath` to lookup the
CommonAppData folder, whose id is 35 (0x23). See the following for
known-folder ids:

  http://source.winehq.org/source/include/shlobj.h

It also updates the acceptance tests to use the `puppetpath` value when
appending the `server` property to the `agent` section of the hosts'
`puppet.conf`. And since the CommonAppData folder contain spaces, the
value of `puppetpath` is quoted on Windows.