(#11767) Add the ability to lazily evaluate a command
Previously, the command_string for a Command was evaluated at command
creation time. As a result, it wasn't possible to specify a command
like:

    on hosts, 'rm -rf #{host["puppetpath"]/ssl}'

and have the command be evaluated in the context of each host.

This commit adds a generic HostCommand that lazily evaluates its
command string, in the `Kernel.eval` sense, with a binding containing
the `host` variable. So now it is possible to do:

    on hosts, host_command('rm -rf #{host["puppetpath"]/ssl')

which will remove the ssl directory from each host using each hosts'
`puppetpath` configuration. Note that we're using single quotes to
avoid string interpolation.