(#11767) Use newer --dns_alt_names option
Previously, the git setup steps were using the old deprecated
--certdnsnames option when generating the master's csr. It was also
using a colon-delimited list instead of a comma-separated list. As a
result, the puppetmasters certificate generated during
03_ValidateSignCert.rb contained bogus X509 Subject Alt Names:

  X509v3 Subject Alternative Name:
    DNS:centos-55-64-1.puppetlabs.lan,
    DNS:puppet:centos-55-64-1:centos-55-64-1

This was not noticed earlier, because the agent never retrieved its
ssl certificate in the setup steps, which is something I recently
changed so that we'd validate that the agent can connect via SSL to
the master in the setup steps, rather than later during the acceptance
tests.

Previously, when running the first acceptance tests from the puppet
repo, the test would delete all of the hosts ssl directories, and
regenerate the certs, but this time using the correct --dns_alt_names
option.

This commit just updates the setup steps to also use the
--dns_alt_names option with a comma delimited list of names to include
in the puppet master's certificate.