Nicer handling of /etc/hosts.
Previously the updates to /etc/hosts would lead to trouble when the machine
wasn't reverted to a snapshot between each test, or if there was already a
hosts entry for that machine.

Now, instead, we carefully clean out the file to get the content we want, and
nothing but, in place for the test suite run.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>