Support PTY allocation on SSH channel.
Previously all remote commands were run without a PTY allocated.  This almost
always delivers the same results as running them with a PTY - but not in every
case.

This adds the ability to request a PTY for a specific test, which will cause
one to be allocated - or a hard failure to occur.

This is not made the default for two reasons:

1. It changes behaviour around which scripts are run on exec.
2. It potentially changes logging behaviour.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>