Clean and forcibly change branch during git setup.
Previously we deleted the entire git repo, which would ensure that we had a
clean, pristine, untouched repository for the next test.

Once we started to reuse the existing checkout we can't assume that - if
someone was to fiddle a file in there, it could break everything.

So, we forcibly clean and checkout now, to restore the repository and checkout
to a pristine state before we proceed through the testing.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>