Checking in one SSH operation rather than one per git command
When we started reusing the repository we needed to do a lot of remote
operations.  While the cost of starting a new process is fairly low, it is at
least a couple of RTTs, which will slow down cloud based testing.

This just rewrites the step to collect the operations and glue them together
into a single operation, cutting that overhead dramatically.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>