(#11944) Refactor command execution logic
Previously, the `TestCase#on` method knew about the inner workings of
the Command class, e.g. that 0 was the default successful exit
code. However, that behavior would only occur if you called the `on`
method, but not if you called the `Command#exec` method directly.

This commit moves the logic into the Command class to ensure the logic
is always applied regardless of how the command is executed.