This comments out all but the UpgradePuppet step in the pe_upgrade dir
This is totally a sledge hammer kind of fix, but the issue is:
CI will be run with snapshots that are ensured to have known good
setups, skipping based on the --vmrun option would be the best for CI

Locally many will be able to use this as is if they also have known
good snapshots.

I also try to automate as much as I can locally testing PE pre-2.0.
The changes that I'd like to introduce to automate those builds are
not trivial and keeping these files is helpful for my work in that area.

NOTE: This works for me locally with a similar (but not exact) workflow
to CI. It *should* work in CI.