(#11944) Add Unix and Windows file commands
This commit defines a set of native commands for validating file
resources on Unix and Windows. This way all of logic for performing
file commands is in one place, rather than scattered throughout each
file test case.

Note the Windows implementation uses cygpath to translate from a posix
path, e.g. /tmp, to a "mixed" windows path, e.g. C:/cygwin/tmp, so
that the latter can be passed to puppet and still be a valid fully
qualified path on Windows.