Allow multiple exitcodes for initial puppet agent run
Prior to this (a change I made to not effectively bounce
the agent service) only accepted an exit code of 0.

This worked for my testing because of conditions that caused
the agent to check in once before the test was actually ran.
The issue is when (and this is the typical case in CI) the test
is the first agent run to check in there will usually be
successful changes applied and puppet will exit with 2.

This adds 2 as an acceptable exit code.

Signed-off-by: Justin Stoller <justin@puppetlabs.com>