Maint: Agent vardir sometimes not deleted between test cases
The `with_master_running_on` method was recently updated to delete the
contents of each agents' vardir so that cached catalogs from previous
runs don't pollute subsequent runs, and cause the agent to return 0
exit code, when it shouldn't, for example.

However, commit `ba57ee5c` refactored this code and caused a
regression, as it was only deleting the agent's vardir if the
directory existed on the host driving the tests, not the system under
test. It also removed the `-f` flag.

This commit reinstates the `-f` flag and moves the `File.exists?`
check so that it is executed on the agent. This is necessary because
calling `for f in /foo/*` returns the literal string '/foo/*' if the
directory foo does not exist.