Maint: Use short, unquoted paths on Windows
Previously, we quoted the puppetpath and puppetvardir host properties
on Windows because they often contain spaces, e.g. C:\Documents and
Settings. However, when these properties are used in bash commands, it
can result in expected consequences. For example,

   if [ -d "#{agent['puppetpath']}" ]

checks if a directory named \"C:\Documents and Settings\...\" exists,
where the quotes are interpreted as part of the path. This was causing
the `with_master_running_on` method to fail to delete Windows agents
vardir contents.

This commit changes the puppetpath and puppetvardir to always return
short versions of Windows paths with forward slashes, e.g.

    C:/DOCUME~1/ALLUSE~1/APPLIC~1/PuppetLabs/puppet/var/client_data

This eliminates the need for these paths to be quoted, and makes it
easier for bash commands to do what you expect in a cross-platform
way.