(Maint) Fix intermittent ssl errors
Previously, when running windows acceptance tests, the ntp step was
sometimes not updating the clock. As a result, the SSL sanity test would
fail due to mismatched clocks.

It turns out the `/update` flag does not necessarily update the client's
time. It just applies the configuration changes, e.g. ntp server IP
address, so that the Windows service can see them. An explicit /resync
is required to force an immediate ntp update. Also the resync occurs
sychronously, because we're not using the `/nowait` switch.