(Maint) Ensure agent certs are signed before proceeding
Previously, there was a race condition whereby "slow agents" may not
submit their CSR before `03_PE-SignCerts` tries to sign them. As a
result, we might not sign anything (since exit code 24 is acceptable for
situations where the master is also an agent and its cert is always
autosigned). We would then get a failure later on in `05_Agent_test`
when validating that the agents could connect to the master.

This commit changes the test to validate that all of the agents'
certificates have been issued before proceeding. Note we only do this
for hosts that are only agents, because the master's cert is autosigned.