(Maint) Add step for installing necessary gems
The FOSS version of Puppet on Windows requires several gems to run. Up
until now, the necessary gems were part of the 'git' snapshot. However,
as we add more features, additional gems will be required. Rather than
update the snapshot each time, this commit adds a setup step that
installs/updates the gems we require. It is written fairly generically
so that platform-specific gems and behavior can be added.

Note this step only occurs in 'git' but not 'pe' because in the latter
case, the gems should be part of the installer itself.