Perform a hard reset to get module changes
Without this patch the --modules command line option results in the
behavior of using git checkout -f on the puppet master systems using the
local branch.

This is a problem because I frequently re-write history in my topic
branches and the upstream branch will no longer be fast-forward-able
relative to the local branch reference.

The end result is that my changes aren't actually tested if I rewrite
history.

This patch fixes the problem by doing a hard reset of the local topic
branch against the upstream reference.  This ensures the local branch is
the exact same as the upstream reference.

The end result is that my changes are tested.  =)