Add modules to Host Configuration Summary
Without this patch the set of modules installed into the puppet master
modulepath is not displayed in the Host Configuration Summary.  This is
a problem because I sometimes forget to switch the branch I'm working on
and I end up running the test suite against a branch that fails the
tests I'm currently working on.

This patch simply adds the array created by --modules to the config hash
object displayed by the summary.

The new summary looks like:

      - Host Configuration Summary -
    Platform for win2003sat windows-2003-amd64
    Platform for qa-ubuntu-lucid ubuntu-10.04-i386
    Role for win2003sat agent
    Role for qa-ubuntu-lucid master
    Role for qa-ubuntu-lucid agent
    Ruby version for win2003sat ruby 1.8.7 (2011-12-28 patchlevel 357) [i386-mingw32]
    Ruby version for qa-ubuntu-lucid ruby 1.8.7 (2010-01-10 patchlevel 249) [i486-linux]
    Config Key|Val: modules ["git@github.com:jeffmccune/puppetlabs-registry.git#maint/master/add_acceptance_tests", "git://github.com/puppetlabs/puppetlabs-stdlib.git#master"]
    Config Key|Val: filecount 12
    Config Key|Val: puppet_ver "git://github.com/puppetlabs/puppet.git#2.7.x"
    Config Key|Val: consoleport 443