Fix false positive error synchronizing win2008r2 Time
Without this patch the acceptance tests fail running against Windows
2008 hosts because the w32tm /register command always returns
"Permission Denied."  Even with UAC completely disabled.

The command is not actually required in Windows 2008.  The commands
following the register command are adequate to synchronize the system
clock after a revert of the VM snapshot.

This change should only affect Windows hosts and the command should
still complete on Windows 2003 hosts where Josh thinks it's required.