Robustify puppet master start/stop
This was using curl to check if the master was stopped, when `kill -0`
will suffice. Also, it was overly permissive in the exit codes it
accepted, and would just wait the full duration for the master to
start/stop in the case where curl wasn't even installed.

Also reduced the timeouts to wait for start/stop, as 30 seconds was a
bit unrealistically long, and just slowed down results when something
really was broken.