Clean-up and stream command output
This merges stdout and stderr, as we've never actually cared much for
the distinction (and this way, they come out in the proper order). Also
reformatted the way output is done to look like a standard shell
session. Command exit codes are only logged if they are non-zero.
Finally, output is now printed as it's received, rather than being
delayed until the command is finished. This provides better insight into
what the test is doing and why it's held up.