Allow options for with_master_running_on
This method now takes an optional final argument, an options hash. The
only current use is to specify the :preserve_ssl option, which dictates
whether or not to leave the ssl directories of the hosts alone. The
default behavior is to remove them, as the master is considered to be a
new puppet instance. However, sometimes a master needs to exist in a
long-running system, so this option can be used to maintain ssl.