Move vmname option from fission-level to host-level
vmname is a generic attribute of a host, thus the option can be useful in
contexts outside of fission. This patch moves that option out of the fission
subsection and into the parent host section.

"I've been wanting to make explicit the distinction between name used by the
harness, hostname, and vmname for awhile. This is something we're going to need
in vSphere soon and I'd like to see 'vmname' as a key in the host hash above
the fission specific section (since I'll probably be using it in vSphere as
well before Lance ships)." -- Justin Stoller

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>