Add "node_name" method to Host class
In certain tests it is very important for us to use
an agent's node_name_value rather than relying
on the hostname from the test config file (or
other possible values that may successfully
resolve in DNS).  For example, when testing
node activation and de-activation for puppetdb,
we *must* use the node_name_value or things
can go awry.

This commit adds a method to the host object
that allows us to retrieve the node_name_value
without having to hard-code the command for
acquiring it into all of the tests.  We might
want to consider similar methods for a few other
things such as FQDN.