Move install_from_git to reusable location
This commit introduces an InstallUtils module (which will likely
become a class, eventually) to house the "install_from_git" method
that used to live in setup/git/01_TestSetup.rb.  By moving this method
to a location under 'lib' instead of 'setup', it can be used during
the custom setup lifecycle phases of individual projects.  Previously
it could only be used implicitly via, e.g., --yagr option, and only
if the --type was set to 'git'.