Add support for loading harness options from a file
This commit introduces a new command-line argument,
"--options-file".  If specified, this file will be
treated as a ruby file to be evaluated, and is
expected to return a hash.  The hash may contain
any of the "normal" options that the framework
supports.  It may also contain extra options
that are specific to the project that is being tested.

Any options passed to the harness on the command
line will be given precedence over the ones found
in the file.