Basic tests for Hosts
This provides tests for most of the host's behaviors, save for `exec`
and `do_scp` which are very similar, but `exec` has a lot of knowledge
in it that should probably be elsewhere. So I've decided to hold off on
spec-ing it until I can refactor its friends at the same time.

Signed-off-by: Justin Stoller <justin@puppetlabs.com>