(#16307) handle solaris specially for bin and sbindir
This is specific to the pkg://solaris/runtime/ruby-18@1.8.7.334,5.11
package available from pkg.oracle.com and available from a full system installation of
solaris11.

The ruby available from the above has Config::CONFIG['bindir'] compiled in to /usr/ruby18/bin
unlike in other platforms and packages which has it set to the system directory /usr/bin
The same issue is present for sbin directory too.

Since puppet and facter need to be delivered to /usr/bin we can not rely on the compiled
in directory for this purpose when using the above package.

This fix updates the invocation of install.rb to pass the directories directly rather than
depend on the compiled in directories. (A better fix would be to rely on the native package
managers to deilver the puppet and facter binaries correctly to the system, so this is a
temporary fix until atleast that happens, or the upstream changes the ruby compiled in
settings in a future version).