Detect missing VMs and fail the setup process
Previously the vSphere integration made a "best effort" pass at the VMs: it
would look for them all, and then act only on the machines that it found.

This means that if a VM name was incorrectly specified or otherwise
unavailable the entire process would just silently skip over it, omitting real
and important test targets.

Now this is detected and a failure results: you can't skip over a VM any longer.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>