Maint: Bump win32-security gem version
Previously, we were using win32-security version 0.1.2, but we need to
bump this to 0.1.4 for puppet 2.7.x and up. The gem was being installed
as a dependency to one of the others.

This commit bumps the gem version. In the future, we will be able to do
`bundle install` (for 3.0.x and up), but am I don't want to break the
acceptance tests right now, as it would require reording this step (to
do the git clone, to get the Gemfile, and then do bundle install).