(maint) Add step to install PE *-test packages
This adds a setup type to install the PE *-test packages. This
leverages the power of the existing PE installer in order to install
the packages correctly on all platforms.

In order to use this you will need a `config/test_packages.yml` file
listing the packages to install as part of this step.