Extra debug logging for master startup
This just adds a bit more verbosity / logging to the output when
the harness is checking to see whether the master is running.  In
cases such as when you have a bad IP address in your hosts file,
this can be tremendously useful in helping to figure out what's
going on.