Only try to stop the master if we have a pidfile
Previously, on rhel6-32-1 I was noticing the test harness hang trying to
execute `bash -c kill $(cat )`, which blocked reading from stdin. The
issue occurred because I reverted the VM and did --no-install. As a
result, the `puppet master --configprint pidfile` command raised an
execption, which flowed through the top-level ensure block.

This commit moves the ensure block so that we only try to stop the
master if we're able to get its pidfile.