During fleet.start retry call fleet.destroy and report on sleep
This addition to the fleet.start retry block will add a fleet.destroy call
before fleet.start is retried in case we need to do some cleanup for it
to work again. It is caught inside its own begin/rescue block and has a
timeout to stop it blocking forever.

I've also added reporting on the number of seconds we are sleeping before
retry.

Signed-off-by: Ken Barber <ken@bob.sh>