(#11276) Install GeoTrust Global CA on Windows
In my local Windows VM, it was not able to connect to the production or
pmtacceptance forge due to the GeoTrust Global CA cert not being present
in my Root certificate store.

This commit adds a `git` setup step that installs the certificate on
Windows. When executed, the step installs the following cert:

  CN=GeoTrust Global CA, O=GeoTrust Inc., C=US

If the cert is already present, the `certutil` command still returns 0,
so no special `acceptable_exit_code` logic is required.